requirements.txt
# --- Environment & HTTP ---
python-dotenv==1.0.1
requests==2.32.5
PyYAML==6.0.2

# --- Data Models & Validation ---
pydantic==2.12.4

# --- NLP / Keyword Extraction ---
yake==0.4.8
markdown-it-py==3.0.0

# --- LangChain / LLM / Agents ---
langchain-core==0.3.15
langchain-google-genai==2.0.1
langgraph==0.2.44

# --- Numeric backend (Python 3.13 compatible) ---
numpy>=2.1.0

# --- Human-in-the-loop (HITL) + Terminal Experience ---
questionary==1.10.0       # interactive prompts for HITL
rich==13.6.0              # beautiful terminal output (tables, markdown, colors)

# --- Reliability / Retry Logic ---
tenacity==9.1.2           # standardized exponential-backoff retries

# --- Optional Utilities (uncomment only if needed) ---
# gitpython==3.1.31        # for cloning & inspecting repos locally
# tqdm==4.67.1             # progress bars for ingestion or multi-repo scanning
# loguru==0.7.0            # advanced logging (optional)